Changed Functional/Internal folder topology
- Added _merge, _mergeN, _props, and _size functions
